error en compras

+++ database/migrations/tenant/2022_08_14_190940_tenant_add_additional_data_to_documents.php
@@ -13,9 +13,11 @@ class TenantAddAdditionalDataToDocuments extends Migration
      */
     public function up()
     {
-        Schema::table('documents', function (Blueprint $table) {
-            $table->json('additional_data')->nullable()->after('additional_information');
-        });
+        if (!Schema::hasColumn('documents', 'state_type_sunat_id')) {
+            Schema::table('documents', function (Blueprint $table) {
+                $table->json('additional_data')->nullable()->after('additional_information');
+            });
+        }
     }
 
     /**
@@ -25,8 +27,10 @@ class TenantAddAdditionalDataToDocuments extends Migration
      */
     public function down()
     {
-        Schema::table('documents', function (Blueprint $table) {
-            $table->dropColumn('additional_data');
-        });
+        if (Schema::hasColumn('documents', 'state_type_sunat_id')) {
+            Schema::table('documents', function (Blueprint $table) {
+                $table->dropColumn('additional_data');
+            });
+        }
     }
 }
